requirements.txt
autopep8==1.6.0
cassandra-driver==3.25.0
click==8.1.3
geomet==0.2.1.post1
pycodestyle==2.8.0
six==1.16.0
toml==0.10.2
